# ToDOv4/api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Task Manager API
  description: This is a simple API
  contact:
    email: joao_nascimento6@hotmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: /
  description: SwaggerHub API Auto Mocking
paths:
...
  /User:
    post:
      tags:
      - User
      summary: Criar User
      description: Criação de um novo User
      operationId: userCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: User criado com sucesso
        "400":
          description: Erro na criação de um User
      x-swagger-router-controller: User
  /User/login:
    get:
      tags:
      - User
      summary: User Login
      description: Inicio de sessão de um User
      operationId: userLogin
      parameters:
      - name: User
        in: query
        description: nome do User
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: password
        in: query
        description: palavra passe do User
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Login efetuado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
        "400":
          description: Parâmetros inválidos. Certifique-se de fornecer todos os valores
            obrigatórios.
          content:
            application/json:
              example:
                error: Parâmetros inválidos. Certifique-se de fornecer todos os valores
                  obrigatórios.
      x-swagger-router-controller: User
  /User/logout:
    get:
      tags:
      - User
      summary: User Logout
      description: Termino de sessão de um User
      operationId: userLogout
      responses:
        "200":
          description: Logout efetuado com sucesso
      security:
      - bearerAuth: []
      x-swagger-router-controller: User
  /User/email:
    get:
      tags:
      - User
      summary: Procurar User apartir do email
      description: Insere o um email valido e encontra o seu proprietario
      operationId: userFindByEmail
      parameters:
      - name: email
        in: query
        description: Necessário enviar o email do User pretendido
        required: true
        style: form
        explode: true
        schema:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
          example: joao@hotmail.com
      responses:
        "200":
          description: Enviado User com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: "Input {email} inválido"
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: User não existe
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: User
    delete:
      tags:
      - User
      summary: Eliminar um user apartir do email
      description: Insere o um email valido e elimina o seu proprietario
      operationId: userDelete
      parameters:
      - name: email
        in: query
        description: Necessário enviar o email do User pretendido
        required: true
        style: form
        explode: true
        schema:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
          example: joao@hotmail.com
      responses:
        "204":
          description: Elimiado com sucesso
        "400":
          description: "Input {email} inválido"
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: User não encontrado
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: User
    patch:
      tags:
      - User
      summary: Atualiza o email do usuário
      description: Inicia sessão e indica o novo email
      operationId: userAtualizarEmail
      parameters:
      - name: email
        in: query
        description: Necessário enviar o email do usuário pretendido
        required: true
        style: form
        explode: true
        schema:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
        example: joao@hotmail.com
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User_email_body'
      responses:
        "201":
          description: Email atualizado com sucesso
        "400":
          description: "Input {email} inválido"
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: User não encontrado.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: User
...
components:
  schemas:
...
    User:
      required:
      - email
      - password
      type: object
      properties:
        nome:
          type: string
        apelido:
          type: string
        email:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
        password:
          type: string
        contacto:
          pattern: "9[0-9]{2}[0-9]{3}[0-9]{3}$"
          type: string
        userId:
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          type: string
          format: uuid
      example:
        password: password
        contacto: "912345678"
        apelido: apelido
        nome: nome
        email: email@hotmail.com
        userId: 550e8400-e29b-41d4-a716-446655440077
...
    Login:
      type: object
      properties:
        email:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
        password:
          type: string
      example:
        email: email@hotmail.com
        password: password
...
    User_email_body:
      type: object
      properties:
        email:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
          example: novoemail@example.com
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT